feat(kilo-vscode): method renaming to improve clarity
