+++ packages/kilo-vscode/src/legacy-migration/sessions/lib/legacy-conversation.ts
